+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.0] - 20201-05-06
+
+* Simplify interface, but also make it more flexible and concise.
+* Decouple `Beacon` type from its value type.
+* Merge multiple find api into `find.beacon`.
+* Allow match beacon by predicate on widget.
+
 ## [0.1.1] - 2021-05-06
 
 * Add `child` to `Beacon`